refactor: moving logic from fly controller to service
