+++ src/api/fly/fly.controller.ts
@@ -1,91 +1,36 @@
-import { FindOptions, RequestContext } from '@mikro-orm/core';
 import type { Request, Response, NextFunction } from 'express';
-import { Fly } from './fly.entity.js';
 import ApiException from '../../core/ApiException.js';
 import { IndexPaginatedEntityResponse } from '../../core/types.js';
-import { Imitatee } from '../imitatee/imitatee.entity.js';
-import { FlyType } from '../flyType/flyType.entity.js';
 import { FlyResourceModel } from './fly.types.js';
-import { mapEntityDbModelToResourceModel } from '../../core/utils.js';
 import { userHasPermission } from '../user/user.service.js';
 import { UserPermissionName } from '../userPermission/userPermission.constants.js';
+import * as flyService from './fly.service.js';
 
 export const indexFlies = async (
     req: Request,
     res: Response<IndexPaginatedEntityResponse<FlyResourceModel>>,
     next: NextFunction,
 ): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(Fly);
-    const pageNumber = Number(req.query.pageNumber);
-    const pageSize = Number(req.query.pageSize);
-    const findOptions: FindOptions<Fly> = {
-        offset: (pageNumber - 1) * pageSize,
-        limit: pageSize,
-        populate: ['types', 'imitatees'] as never,
-        orderBy: { name: 'ASC' },
-    };
-
-    const results = await repository?.findAndCount({}, findOptions);
-
-    if (!results) {
-        const error = new ApiException({ message: 'Unable to fetch flies' });
-        return next(error);
+    try {
+        const result = await flyService.indexFlies(Number(req.query.pageNumber), Number(req.query.pageSize));
+        res.json(result);
+    } catch (e) {
+        next(e);
     }
-
-    const flies = results[0];
-    const totalItems = results[1];
-    const totalPages = Math.ceil(totalItems / pageSize);
-
-    const mappedResults: Array<FlyResourceModel> = flies.map((x) => {
-        const { externalId: _externalId, id: _id, ...entityModel } = x;
-        return {
-            ...entityModel,
-            id: x.externalId,
-            imitatees: x.imitatees.toArray().map(mapEntityDbModelToResourceModel),
-            types: x.types.toArray().map(mapEntityDbModelToResourceModel),
-        };
-    });
-
-    res.setHeader('Content-Range', `bytes 0-${totalItems}/*`);
-    res.json({
-        metadata: {
-            totalItems,
-            pageNumber,
-            pageSize,
-            totalPages,
-        },
-        results: mappedResults,
-    });
 };
 
 export const getFly = async (req: Request, res: Response<FlyResourceModel>, next: NextFunction): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(Fly);
-    const result = await repository?.findOne({ externalId: req.params.id }, { populate: ['types', 'imitatees'] });
-
-    if (!result) {
-        const error = new ApiException({ message: `Cannot find fly using id: ${req.params.id}`, status: 409 });
-        return next(error);
+    try {
+        const result = await flyService.getFly(req.params.id);
+        res.json(result);
+    } catch (e) {
+        next(e);
     }
-
-    const { externalId: _externalId, id: _id, ...mappedResult } = result;
-    const response: FlyResourceModel = {
-        ...mappedResult,
-        id: result.externalId,
-        imitatees: result.imitatees.toArray().map(mapEntityDbModelToResourceModel),
-        types: result.types.toArray().map(mapEntityDbModelToResourceModel),
-    };
-
-    res.json(response);
 };
 
 export const createFly = async (req: Request, res: Response<string>, next: NextFunction): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(Fly);
-    const flyTypeRepository = em?.getRepository(FlyType);
-    const imitateeRepository = em?.getRepository(Imitatee);
     const userId = req.body.user.userId;
+
     const hasPermission = await userHasPermission(userId, UserPermissionName.CREATE);
 
     if (!hasPermission) {
@@ -96,60 +41,15 @@ export const createFly = async (req: Request, res: Response<string>, next: NextF
         return next(error);
     }
 
-    const exists = await repository?.exists(req.body.name);
-
-    if (exists) {
-        const error = new ApiException({ message: 'Fly name already exists', status: 409 });
-        return next(error);
+    try {
+        const externalId: string = await flyService.createFly(req.body);
+        res.json(externalId);
+    } catch (e) {
+        next(e);
     }
-
-    const fly = new Fly(req.body.name, req.body.description);
-
-    const flyTypes: Array<FlyType> = [];
-    const imitatees: Array<Imitatee> = [];
-
-    for (let i = 0; i < req.body.types.length; i++) {
-        const id = req.body.types[i];
-        const flyType = await flyTypeRepository?.findOne({ externalId: id });
-        if (!flyType) {
-            const error = new ApiException({ message: `Cannot find fly type using ID ${id}` });
-            return next(error);
-        }
-        const reference = flyTypeRepository?.getReference(flyType.id);
-        if (!reference) {
-            const error = new ApiException({ message: `Cannot find fly type using ID ${id}` });
-            return next(error);
-        }
-        flyTypes.push(reference);
-    }
-
-    for (let i = 0; i < req.body.imitatees.length; i++) {
-        const id = req.body.imitatees[i];
-        const imitatee = await imitateeRepository?.findOne({ externalId: id });
-        if (!imitatee) {
-            const error = new ApiException({ message: `Cannot find imitatee using ID ${id}` });
-            return next(error);
-        }
-        const reference = imitateeRepository?.getReference(imitatee.id);
-        if (!reference) {
-            const error = new ApiException({ message: `Cannot find imitatee using ID ${id}` });
-            return next(error);
-        }
-        imitatees.push(reference);
-    }
-
-    fly.types.add(flyTypes);
-    fly.imitatees.add(imitatees);
-
-    await em?.persist(fly).flush();
-    res.json(fly.externalId);
 };
 
 export const updateFly = async (req: Request, res: Response<FlyResourceModel>, next: NextFunction): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(Fly);
-    const flyTypeRepository = em?.getRepository(FlyType);
-    const imitateeRepository = em?.getRepository(Imitatee);
     const userId = req.body.user.userId;
     const hasPermission = await userHasPermission(userId, UserPermissionName.CREATE);
 
@@ -161,70 +61,15 @@ export const updateFly = async (req: Request, res: Response<FlyResourceModel>, n
         return next(error);
     }
 
-    const fly = await repository?.findOne({ externalId: req.params.id }, { populate: ['types', 'imitatees'] });
-
-    if (!fly) {
-        const error = new ApiException({ message: `Cannot find fly using id: ${req.params.id}`, status: 409 });
-        return next(error);
+    try {
+        const fly = await flyService.updateFly(req.body);
+        res.json(fly);
+    } catch (e) {
+        next(e);
     }
-
-    const { name, description } = req.body;
-
-    fly.name = name;
-    fly.description = description;
-
-    const flyTypes: Array<FlyType> = [];
-    const imitatees: Array<Imitatee> = [];
-
-    for (let i = 0; i < req.body.types.length; i++) {
-        const id = req.body.types[i];
-        const flyType = await flyTypeRepository?.findOne({ externalId: id });
-        if (!flyType) {
-            const error = new ApiException({ message: `Cannot find fly type using ID ${id}` });
-            return next(error);
-        }
-        const reference = flyTypeRepository?.getReference(flyType.id);
-        if (!reference) {
-            const error = new ApiException({ message: `Cannot find fly type using ID ${id}` });
-            return next(error);
-        }
-        flyTypes.push(reference);
-    }
-
-    for (let i = 0; i < req.body.imitatees.length; i++) {
-        const id = req.body.imitatees[i];
-        const imitatee = await imitateeRepository?.findOne({ externalId: id });
-        if (!imitatee) {
-            const error = new ApiException({ message: `Cannot find imitatee using ID ${id}` });
-            return next(error);
-        }
-        const reference = imitateeRepository?.getReference(imitatee.id);
-        if (!reference) {
-            const error = new ApiException({ message: `Cannot find imitatee using ID ${id}` });
-            return next(error);
-        }
-        imitatees.push(reference);
-    }
-
-    fly.types.set(flyTypes);
-    fly.imitatees.set(imitatees);
-
-    await em?.flush();
-
-    const { externalId: _externalId, id: _id, ...mappedResult } = fly;
-    const response: FlyResourceModel = {
-        ...mappedResult,
-        id: fly.externalId,
-        imitatees: fly.imitatees.toArray().map(mapEntityDbModelToResourceModel),
-        types: fly.types.toArray().map(mapEntityDbModelToResourceModel),
-    };
-
-    res.json(response);
 };
 
 export const deleteFly = async (req: Request, res: Response<string>, next: NextFunction): Promise<void> => {
-    const em = RequestContext.getEntityManager();
-    const repository = em?.getRepository(Fly);
     const userId = req.body.user.userId;
     const hasPermission = await userHasPermission(userId, UserPermissionName.DELETE);
 
@@ -236,12 +81,10 @@ export const deleteFly = async (req: Request, res: Response<string>, next: NextF
         return next(error);
     }
 
-    const result = await repository?.nativeDelete({ externalId: req.params.id });
-
-    if (result === 0) {
-        const error = new ApiException({ message: `Cannot find fly using id: ${req.params.id}` });
-        return next(error);
+    try {
+        await flyService.deleteFly(req.body.id);
+        res.json('OK');
+    } catch (e) {
+        next(e);
     }
-
-    res.json('OK');
 };
